workaround cyclic dependency issue
