+++ src/vs/workbench/api/common/extHostTypeConverters.ts
@@ -1299,23 +1299,7 @@ export namespace LogLevel {
 
 export namespace NotebookCellOutput {
 	export function from(output: types.NotebookCellOutput): IDisplayOutput {
-
-		let data: { [key: string]: unknown; } = {};
-		let custom: { [key: string]: unknown; } = {};
-		let hasMetadata = false;
-
-		for (let item of output.outputs) {
-			data[item.mime] = item.value;
-			if (item.metadata) {
-				custom[item.mime] = item.metadata;
-				hasMetadata = true;
-			}
-		}
-		return {
-			outputKind: CellOutputKind.Rich,
-			data,
-			metadata: hasMetadata ? { custom } : undefined
-		};
+		return output.toJSON();
 	}
 }
